Nullify os library to make FCEUX safer.
Nullfied the 'os' library included in lua by default, making FCEUX as a whole safer, but still not foolproof. For my own project, I require the io class to still be intact, which unfortunately means malicious programs can still get through to some degree. If someone wants to make a fork of this that's even lighter (a FCEUX-Zero if you will), then be my guest!